% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carFunctions.R
\name{g}
\alias{g}
\title{A function to calculate G in the Pocock Design and return a vector of p}
\usage{
g(x, p1 = 3/4, n.trt)
}
\arguments{
\item{x}{x is the vector of deltas in Pocock's design}

\item{p1}{= 3/4}
}
\description{
This function is not generally used. It is used in the calculation of the treatment assignments in psd
}
