% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genCovValues.R
\name{genCovValues}
\alias{genCovValues}
\title{Generate covariate Values}
\usage{
genCovValues(p = c(0.5, 0.5), N = 200)
}
\arguments{
\item{N}{The total sample size}

\item{p1}{The probability vector for covariate 1}

\item{p2}{The probability vector for in covariate 2}
}
\description{
This is a function generate covariate values to use before an allocation scheme
}
\examples{
genCovValues()
}
