% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatisticFunctions.R
\name{get.t.one}
\alias{get.t.one}
\title{One Treatment Test Statistic}
\usage{
get.t.one(mu.gd.vec, var.mu.gd.vec, z.alpha, mu.null)
}
\arguments{
\item{mu.gd.vec}{This is the vector of the means of weighted means}

\item{var.mu.gd.vec}{This is the vector of the variance of weighted means}

\item{z.alpha}{This is the set alpha level.}

\item{mu.null}{Mu.null is the mu under the null distribution}
}
\description{
This function finds the test statistic for a two-treatment case of Chow & Tse (2007)
}
