% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testStatisticFunctions.R
\name{get.t}
\alias{get.t}
\title{get.t}
\usage{
get.t(beta0, beta1, mu.null = 0, sigma = NULL, tau = NULL,
  z.alpha = 0.05, z.beta = 0.2)
}
\arguments{
\item{beta0}{Beta0 is the beta_0 coefficient from the known relationship between the Y and X variables}

\item{beta1}{Beta1 is the beta_1 coefficient from the}

\item{sigma}{Sigma is the known variance}

\item{tau}{Tau is the known tau from the weight relationship}

\item{data}{Data is the dataset containing short-term and long-term endpoints}

\item{alpha}{Alpha is the Type I error rate level}
}
\description{
This function gets the test statistic for comparing a treatment against a mean
}
