% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carFunctions.R
\name{pbr}
\alias{pbr}
\title{PBR Function}
\usage{
pbr(n, block.size = 4)
}
\arguments{
\item{n}{Total n size}

\item{block.size}{This is the number of subjects in each block}
}
\description{
This function is not generally used. This is used in spbd to determine treatment assignments
}
