% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carFunctions.R
\name{psd}
\alias{psd}
\title{Function for returning treatment assignments in Pocock's Design}
\usage{
psd(x, p1 = 3/4, best = 0, tr = NULL, n.trt)
}
\arguments{
\item{x}{Vector of factor level assignments}

\item{tr}{Tr is the vector of treamtnet assignments. Default is NULL}

\item{n.trt}{Number of treatments used in the trial}

\item{p}{Default is 3/4.}

\item{Best}{is the selected best treatment if past the first analysis. Default is 0}

\item{n.trt}{The number of treatments used in the analaysis if at the first stage in the analysis.}
}
\description{
This function takes a vector of factor level assignments, and returns treatment assignments
}
\examples{
psd(x = seamlessTrials::all)
}
