% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/carFunctions.R
\name{spbd}
\alias{spbd}
\title{spbd Function}
\usage{
spbd(x, n, m = 4)
}
\arguments{
\item{x}{X vector of covariate values, ranging from 0 to total number of factor levels}

\item{n}{Total sample size}

\item{m}{The number of blocks}
}
\description{
Function to determine treatment assignments in SPB Design
}
\examples{
spbd()
}
